Revert "Loan.java File got modified"
